==> prep/bio/spp/1_get_ne_iucn_spp.Rmd <==
---
title: 'OHI-Northeast: 1. Get list of IUCN Species maps in the Northeast'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

# Summary

This script takes the OHI Northeast region shapefile and compares it to the global shapefiles we have from IUCN to extract all species maps that fall within our region. A lot of this code dips into the `spp_risk_dists` repo from Casey O'Hara (forked to my repo): https://github.com/jamiecmontgomery/spp_risk_dists.

# Setup
```{r setup, message = F, warning = F}
knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)
source('~/github/ne-prep/src/R/common.R')

library(sf)
library(raster)
library(rfishbase)
library(rredlist)
library(taxize)

#ne region shapefile
ne_shp <- rgns_simp
```

I need to get a list of the cell IDs for the Northeast region. I'm using the global cell ID raster from Casey O'Hara's `spp_risk_dists` repository. 

```{r global_cells}
#these are global cell ids in raster format pulled from the spp_risk_dists repo.
cells <- raster("~/github/spp_risk_dists/_spatial/cell_id_rast.tif")
```
Then I select only those cells in the Northeast and use that list to query the larger species data. First we transform our regions shapefile into the same CRS as the cell ID raster (`cells`).

```{r reproject_rgns}
#reproject to our CRS
ne_reproj <- st_transform(ne_shp, crs = "+proj=cea +lon_0=0 +lat_ts=45 +x_0=0 +y_0=0 +ellps=WGS84
+units=m +no_defs")

plot(cells, col = cols)
plot(ne_reproj, add = T)
```

Extract the cell ids for our region
```{r extract_ne_cell_ids}
ne_cells <- raster::extract(cells, ne_reproj) %>%
  unlist()
```

Read in the necessary files to link species to their map's filepath and species info
```{r spp_files}
# .csv file that lists all species and the file path to their map
spp_maps <- read_csv('~/github/spp_risk_dists/_data/spp_marine_maps_2018-1.csv',
                     col_types = 'ddciccc')

#file with species information to link to iucn_sid at the end
spp_info <- read_csv("/home/ohara/git-annex/spp_risk_dists/iucn/spp_info_from_api_2018-1.csv")
```

The following forloop goes through each species:  
- finds it's species path
- reads in the species .csv that lists all global cells where it is found
- filters that list to just those in the Northeast
- returns an aggregated dataframe (`taxa_cells_df`) with all species scientific names, their unique ids (sids), and cellIDs

```{r, eval = F}
#grab each taxa's folder filepath
taxa <- spp_maps$dbf_file %>%
    unique() %>%
    str_replace('\\....$', '')

#create an empty list that is the length of all taxa. We are going to fill this list  
taxa_cells_list <- vector('list', length = length(taxa))

#for each taxa, grab the species map (raster) and filter to only keep those cells in the northeast.
for(i in seq_along(taxa)) { ### i <- 5
    taxon <- taxa[i]
    print(i)
    spp_ids_in_taxon <- spp_maps %>%
      filter(str_detect(dbf_file, taxon)) %>%
      .$iucn_sid
    cat(sprintf('processing %s spp in %s...\n', length(spp_ids_in_taxon), taxon))
    
    spp_cells <- parallel::mclapply(spp_ids_in_taxon, mc.cores = 32,
                                    FUN = function(x) { ### x <- spp_ids_in_taxon[1]
                                      f <- file.path('/home/ohara/git-annex/spp_risk_dists/spp_rasters',
                                                     sprintf('iucn_sid_%s.csv', x))
                                      if(file.exists(f)) {
                                        y <- read_csv(f, col_types = 'di') %>%
                                          mutate(iucn_sid = x) %>%
                                          select(-presence)  %>%
                                          filter(cell_id %in% ne_cells)
                                      } else {
                                        y <- data.frame(cell_id = NA,
                                                        iucn_sid = x, 
                                                        f = f, error = 'file not found')
                                      }
                                      return(y)
                                    }) %>%
      bind_rows() %>%
      mutate(spp_gp = taxon)
    
    taxa_cells_list[[i]] <- spp_cells
}
  
taxa_cells_df <- taxa_cells_list %>%
    bind_rows()  %>%
    filter(!is.na(cell_id)) %>%
    select(iucn_sid) %>%
    distinct() %>%
    left_join(spp_info)
```

We don't have any common names for these maps! So we're using `rredlist` to get common names

```{r}
scinames <- taxa_cells_df$sciname

iucn_common_names <- data.frame()
for(i in 1:length(scinames)){
  sp <- scinames[i]
  print(i)

  if(sp %in% c("Phalacrocorax auritus", "Raja eglanteria")){ #this one threw some error and no $result was returned from rl_search
    comm <- NA
  }else{
  comm <- rredlist::rl_search(sp)$result$main_common_name
  }
  
  df <- data.frame(sciname = sp,
                   common = comm)
  
  iucn_common_names <- rbind(iucn_common_names, df) %>%
    mutate(common = as.character(common),
         sciname = as.character(sciname))
}
```


We are still missing `r nrow(filter(iucn_common_names, is.na(common)))`. Let's try fishbase?


```{r, eval = F}
sp <- iucn_common_names %>%
  filter(is.na(common))  %>%
  select(sciname)

fb_list <- species(sp$sciname) %>%
  select(Species, FBname)
```

This got us `r nrow(filter(fb_list, !is.na(FBname)))` more common names we need to add back in.

```{r}
#add back in common names
out <- iucn_common_names %>%
  left_join(fb_list, by = c("sciname" = "Species")) %>%
  mutate(common = ifelse(is.na(common), FBname, common)) %>%
  select(-FBname)
```

Ok we got a lot of them but not all. Now we can try to use the `taxize` R package.

```{r, eval = F}
missing_scinames <- out %>%
  filter(is.na(common)) %>%
  .$sciname

common_names <- c()
for(i in 1:length(missing_scinames)){ #i <- 1
  sp <- missing_scinames[i]
  comm <- taxize::sci2comm(sp, db = "itis")
  print(comm)
common_names <- c(common_names, comm)
}

#turn list into dataframe
df <- do.call(rbind,lapply(common_names,data.frame))
df$sciname <- rownames(df)

df2 <- df %>%
  rename(common = X..i..) %>%
  filter(!is.na(common)) %>%
  mutate(sciname = str_extract(sciname, "[^.]+")) #remove numbers
```

Only 4 of the species had a common name via taxize. We will add these manually and then the remaining species will keep NA for common name.

```{r, eval = F}
sci_to_comm <- out %>%
  mutate(common_name = 
           case_when(
              sciname == "Lestrolepis intermedia" ~ "Barracudina antifaz",
              sciname == "Astrangia poculata" ~ "Northern star coral",
              sciname == "Gigantactis vanhoeffeni" ~ "Cosmopolitan whipnose",
              sciname == "Phalacrocorax auritus" ~ "Double-crested cormorant",
              TRUE ~ as.character(common))) %>%
  select(-common)
```

```{r}
iucn_spp_info <- taxa_cells_df %>%
  select(iucn_sid, sciname, population, category) %>%
  left_join(sci_to_comm)
```


```{r, eval = F}
#save
write_csv(iucn_spp_info, "data/1_iucn_spp_in_ne.csv")
```

Filter the list of IUCN marine maps from Casey's repo to just have the file paths for species in the Northeast
```{r, eval = F}
spp_marine_maps <- read_csv("~/github/spp_risk_dists/_data/spp_marine_maps_2018-1.csv") %>%
  filter(sciname %in% iucn_spp_info$sciname,
         subpop %in% c("Atlantic Northwest", "Northwest Atlantic", NA)) #all other subpops that have regions are not in the Northeast so remove here
#save
write.csv(spp_marine_maps, "data/1_iucn_spp_shp_filepaths.csv")
```
