==> prep/bio/spp/2_rasterize_iucn_spp_shps.Rmd <==
---
title: 'Rasterize IUCN species maps'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

# Summary

Using a set of IUCN species range maps, rasterize each species range map to 1km x 1km raster using `fasterize`.  Use `presence` field from shapefile.

* Subpopulation polygons must be identified and rasterized separately from the parent polygon; this must be done by sciname and subpop fields since the polygon IDs are based upon the parent ID.
* Regional assessments need not be determined at this stage - the ID numbers match the global ID numbers (including subpops).

# Data source

IUCN Red List: Spatial Data Download
IUCN: Casey O'Hara via Gina Ralph direct communication

# Setup 

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/', message = FALSE, warning = FALSE)

library(raster)
library(rgeos)
source('~/github/ne-prep/src/R/common.R')
library(sf)


dir_bli <- file.path(dir_M, 'git-annex/globalprep/_raw_data/birdlife_intl/d2018')
dir_shp <- file.path(dir_M, 'git-annex/globalprep/_raw_data/iucn_spp/d2018-1')
  ### These are shapefiles directly from IUCN as individual species map files
  ### (in the unzipped folder).
  ### in this folder are shapefiles at a taxonomic level.
```

# Methods

## Read spp shapes, correct subpop IDs, `fasterize()`, depth clip, save to csv

We will loop over each species in each shapefile and rasterize separately, using `sf` and `fasterize` packages.  

* From the full map list, filter to a single shapefile
* Load shapefile using `st_read`, and correct subpop IDs from `shp_iucn_sid` to `iucn_sid`
* Loop over each `iucn_sid` in the shapefile, rasterizing (`fasterize()`) to 10 km^2 resolution, using "presence" field. 
    * clip to neritic (<=200 m) and shallow (<=60 m) depth raster if appropriate.  Otherwise mask to bathy raster.  Since bathy raster was created by masking to area raster, cells with any marine presence will be kept but any non-marine cells will be dropped.
    * Save as .tif and .csv, and compare average file sizes.  .csv easier to work with, but might be significantly larger than .tifs in the long run.
        * note: no longer saving as .tif for speed and file size - just use .csv instead!
    * use `mclapply()` to speed this up.
    
``` {r rasterize and clip and save to csv}

maps_to_rasterize <- read_csv("data/1_iucn_spp_shp_filepaths.csv") %>%
  select(-X1) %>%
  mutate(shp_file = str_replace(dbf_file, 'dbf$', 'shp'))

### rast_base for cell IDs
rast_base <- raster("~/github/ne-prep/spatial/ocean_rasters/ne_rgns_rast.tif") %>%
  projectRaster(crs = us_alb)

  
################################################################.
### Loop over each distinct shapefile with species range maps
################################################################.
  shps <- maps_to_rasterize$shp_file %>% unique()	
  for(i in seq_along(shps)) {	
    ### i <- 1	
    	
    shp <- shps[i]	
    	
    maps_in_shp <- maps_to_rasterize %>%	
      filter(shp_file == shp)	
    	
    id_fix <- maps_in_shp %>%	
      select(shp_iucn_sid, iucn_sid, subpop, max_depth) %>%	
      distinct()	
    	
    polys_all <- read_sf(shp, type = 6)	
      ### we will check geoms and fix them inside the mclapply, 	
      ### and then reproject, one species at a time	
    if(!'sciname' %in% names(polys_all)) {	
      polys_all <- polys_all %>%	
        rename(sciname = binomial)	
    }	
    if(!'subpop' %in% names(polys_all)) {	
      polys_all$subpop <- NA_character_	
      ### if shape doesn't have subpop column, add it as NA	
    }	
    if('id_no' %in% names(polys_all)) {	
      polys_all <- polys_all %>%	
        rename(iucn_sid = id_no)	
    }	
    if(!'presence' %in% names(polys_all)) {	
      polys_all <- polys_all %>%	
        mutate(presence = 1)	
    }	
    	
    polys_match <- polys_all %>%	
      select(shp_iucn_sid = iucn_sid, sciname, subpop, presence, geometry) %>%	
      mutate(presence = ifelse(presence == 0, 1, presence),	
             subpop   = as.character(subpop)) %>%	
      inner_join(id_fix, by = c('shp_iucn_sid', 'subpop')) 	
    	
    spp_ids <- maps_in_shp$iucn_sid %>% 	
      sort() %>% 	
      unique()	
    	
  ### These will be used as masks	
  rast_bathy <- raster('~/github/spp_risk_dists/_spatial/bathy_rast.tif') %>%	
    projectRaster(crs = us_alb)	
  rast_neritic <- raster('~/github/spp_risk_dists/_spatial/bathy_rast_neritic.tif') %>%	
    projectRaster(crs = us_alb)	
  rast_shallow <- raster('~/github/spp_risk_dists/_spatial/bathy_rast_shallow.tif') %>%	
    projectRaster(crs = us_alb)	
    ####################################################################.	
    ### In each shapefile, loop over each species ID using mclapply().	
    ####################################################################.	
    	
      tmp <- parallel::mclapply(seq_along(spp_ids),	
                                mc.cores = 12, 	
                                FUN = function(x) {	
        ### x <- 1	
        spp <- spp_ids[x]	
        	
        spp_shp <- polys_match %>%	
          filter(iucn_sid == spp)	
        	
       # spp_shp <- valid_check(spp_shp)	
          ### if invalid geom, and bounds exceeded, buffer to 0	
        spp_shp <- spp_shp %>%	
          st_transform(crs = us_alb)	
        	
        spp_rast <- fasterize::fasterize(spp_shp, rast_base, field = 'presence', fun = 'min')	
        	
        ### depth clip if necessary; otherwise clip to bathy raster (which previously	
        ### was clipped to area raster - so cells with any marine area will be kept,	
        ### and non-marine cells will be dropped)	
        if(spp_shp$max_depth == '< 20 m') {	
          spp_rast <- mask(spp_rast, rast_shallow)	
        } else if(spp_shp$max_depth == '< 200 m') {	
          spp_rast <- mask(spp_rast, rast_neritic)	
        } else {	
          spp_rast <- mask(spp_rast, rast_bathy)	
        }	
        ### write out as a raster:	
        rast_file <- file.path(dir_anx, sprintf('bio/spp_presence_rasters/iucn_sid_%s.tif', spp))	
        raster::writeRaster(spp_rast, rast_file, overwrite = TRUE)	
        return(NA)	
      }) ### end of mclapply FUN definition	
    # }) ### end of system.time call	
  } ### end of for loop over each species group

```
