==> prep/bio/spp/5_create_spp_source_status_df.Rmd <==
---
title: "Create info table for all species"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

# Summary

This script creates a dataframe that contains all species in the Northeast for which we have maps either from IUCN RedList or the Northeast Ocean Data Portal. It also links each species with the status based on NatureServe information pulled from `4_get_natureserve_data.Rmd` or IUCN status information from the IUCN API.

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
source('~/github/ne-prep/src/R/common.R')
```

# Get data

Get list of the species we are using, from both IUCN and the NE Data Portal. Also get the species status

```{r}
#ne dataportal rasters
ne_rasters <- list.files(file.path(dir_anx, "bio/portal_spp_rasters"))

#iucn rasters
iucn_rasters <- list.files(file.path(dir_anx, "bio/spp_presence_rasters"), full.names = T)
```

The names in the `iucn_rasters` list only have the species ID. So we need to figure out what species we have in the IUCN list that is also in the data portal list by linking sids.

```{r}
#list of iucn species in northeast. This containts the iucn_sid column.
iucn_ne_spp <- read_csv("data/1_iucn_spp_in_ne.csv") %>%
  select(iucn_sid, common_name, sciname) %>%
  mutate(common = tolower(common_name))
```

Get data portal species

```{r}
ne_dp_spp <- read_csv("data/4_ne_dataportal_spp_common_sci.csv") %>%
  mutate(source = "dataportal")
```

Join the IUCN and NE to better identify which species are in both so we can prioritize the NE data portal maps for those.
```{r}
j <- iucn_ne_spp %>%
  left_join(ne_dp_spp, by = "common") %>%
  left_join(ne_dp_spp, by = c("sciname.x" = "sciname")) %>%
  mutate(common = ifelse(is.na(common.y), common.x, common.y),
         sciname = ifelse(is.na(sciname.y), sciname.x, sciname.y),
         source = ifelse(is.na(source.y), "IUCN", "dataportal")) %>%
  select(iucn_sid, common, sciname, source)
```

There are `r nrow(j%>%filter(source == "dataportal"))` data portal species that are also in the IUCN dataset. 

```{r}
species_df <- ne_dp_spp %>%
  mutate(iucn_sid = NA) %>%
  bind_rows(j) %>%
  mutate(keep = ifelse(!is.na(iucn_sid) & source == "dataportal", 0, 1)) %>% #we now have duplicate rows for data portal species, one with an NA for iucn_sid and one with
  filter(keep == 1) %>%
  select(-keep) %>%
  distinct()
```

The total number of species is `r nrow(species_df)`

Next step is to create a large dataframe that lists each species, what map source we will use for the species, the status and the geographic location of that status (e.g. some states have different statuses for the same species)

```{r}
spp_status    <- read_csv("data/4_natureserve_spp_status.csv") %>% select(-X1)

large_df <- species_df %>%
  left_join(spp_status, by = c("sciname" = "species"))
```

How many of our species have a status assessment?

```{r}
large_df %>%
  filter(!is.na(status)) %>%
  select(common, sciname) %>%
  distinct() %>%
  nrow()
```

Eek so the majority of our species do not have a status. I'm going to try and get more status information from the IUCN API for IUCN species.

```{r, eval = F}
#get iucn species without a status
iucn_sp_wo_status <- large_df %>%
  filter(source == "IUCN",
         is.na(status)) %>%
  select(common, sciname, iucn_sid) %>%
  distinct()

#load the library that accesses the IUCN API
library(rredlist)

#forloop for each species to grab status and trend

df_iucn <- data.frame()

for(i in 1:nrow(iucn_sp_wo_status)){
  
  sp <- as.character(iucn_sp_wo_status[i,2]) #grab scientific name

  cat <- rl_search(sp)$result$category
  tr <- rl_search(sp)$result$population_trend
  
  df2 <- data.frame(sciname = sp,
                   status = cat,
                   trend = tr)
  
  df_iucn <- rbind(df_iucn, df2)
}
```

I'll do the same, using the IUCN API, for Northeast dataportal species without a status. This is just `r large_df %>% filter(source == "dataportal", is.na(status)) %>% nrow()` species.

```{r}
#get ne_dp species without a status
ne_dp_wo_status <- large_df %>%
  filter(source == "dataportal",
         is.na(status))

#function to get category and trend
f <- function(sp){
  cat <- rl_search(sp)$result$category
  tr  <- rl_search(sp)$result$trend
  df <- data.frame(sciname = sp,
                   status = cat,
                   trend = ifelse(is.null(tr), NA, tr))
  return(df)
}

#apply function to each species, result is a dataframe of all species and their category/trend.
df_np <- data.frame()

for(i in 1:nrow(ne_dp_wo_status)){ #
  
  sp <- as.character(ne_dp_wo_status[i, 2]) #grab scientific name
  #print(sp)

  possibleError <- tryCatch(
      df <- f(sp),
      error=function(e) e)
if(inherits(possibleError, "error")) next
  
  df_np <- rbind(df_np, df)
  
}

df_np
```

Looks like these species don't have a status in the IUCN database!

Add status category back to `large_df`

```{r}
d <- large_df %>%
  left_join(df_iucn, by = "sciname") %>%
  rename(status = status.x,
         iucn_status = status.y) %>%
  mutate(status = ifelse(is.na(status), as.character(iucn_status), status), #use IUCN status for IUCN species without a status
         status_scale = ifelse(!is.na(status) & is.na(state) & source == "IUCN", "IUCN", state)) %>%
  select(-iucn_status, -X1)  #remove the interim IUCN status column 
```

How many species now have a status? Remove data deficient

```{r}
d %>%
  filter(!is.na(status),
         status != "DD",
         status !="DD - Data deficient") %>%
  select(sciname) %>%
  distinct() %>%
  nrow()
```

Save

```{r}
write.csv(d, file = "data/5_spp_status.csv")
```
