==> prep/bio/spp/6_link_spp_cells_rgns.Rmd <==
---
title: "Linking species maps to cell ids"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---
# Summary

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
source('~/github/ne-prep/src/R/common.R')
```


## Where do these species exist within our region?

To identify where these species are, based on the range maps we have, we need to do an overlay of our Northeast regions and each map.

```{r}
#this raster is each rgn 1-11 rasterized at the 1km2 cell level
nergns_rast <- raster("~/github/ne-prep/spatial/ocean_rasters/ne_rgns_rast.tif")

#this raster has individual cell id values
cellid_rast <- raster("~/github/ne-prep/spatial/ocean_rasters/ne_cellids.tif")
  
#dataframe that links cellIDs with the OHI region they are in
rgn_cells <- read.csv("~/github/ne-prep/spatial/rgn_cellids.csv") %>%
  select(-X)
```

Get IUCN SID's for the species that we are not going to use the IUCN maps for (ones that are also in the dataportal).
```{r}
iucn_sids <- read_csv("data/5_spp_status.csv") %>%
  filter(source == "IUCN") %>%
  select(iucn_sid) %>%
  distinct() %>%
  .$iucn_sid
```

List of dataportal rasters

```{r}
portal_list <- list.files(file.path(dir_anx, "bio/portal_spp_rasters"), full.names = T)
```

Get regions where species is present. This is not by area - if a species range has 1 cell or 1000 cells in a region, it is counted the same.
```{r, eval = F}
get_sp_rgns <- function(spp){
  
  #get SID (species ID)
  sp <- substr(spp, 58, nchar(spp)-4)
  
  #read in raster
  r <- raster(spp)
  
  #use the raster::zonal() function to get a dataframe that lists each region where the species is found and a column with the species name
  z <- zonal(r, nergns_rast, na.rm=T, fun = "count") %>%
    as.data.frame() %>%
    rename(rgn_id = zone) %>%
    mutate(species = sp) %>%
    filter(count > 0) %>% 
    select(-count)

}

dp_spp_rgns <- purrr::map_df(portal_list, get_sp_rgns) 

#take list of all species and add rgn_id 12

rgn12 <- dp_spp_rgns %>%
  select(species) %>%
  distinct() %>%
  mutate(rgn_id = 12)

dp_spp_rgns_ne <- bind_rows(rgn12, dp_spp_rgns)

write.csv(dp_spp_rgns_ne, file = "data/6_ne_dataportal_spp_by_rgn.csv")
```

We also want to know exactly what cells each species is in. This will primarily be used to create a final map that shows all species risk in the Northeast.

```{r, eval = F}
dp_cells_extract <- function(spp){
  
  #get SID (species ID)
  sp <- substr(spp, 58, nchar(spp)-4)
  
  #read in raster
  r <- raster(spp)
  
  #use the raster::zonal() function to get a dataframe that contains cellids and SID as two columns
  z <- zonal(r, cellid_rast, na.rm=T) %>%
    as.data.frame() %>%
    filter(!is.na(mean)) %>%
    rename(cellID = zone) %>%
    mutate(species = sp) %>%
    inner_join(rgn_cells) %>% #innerjoin because I only want cells that fall inside our region boundaries
    select(-mean)
}
dp_cells <- purrr::map_df(portal_list, dp_cells_extract)

#save to mazu because of file size
write.csv(dp_cells, file = file.path(dir_anx, "bio/ne_dataportal_spp_cells.csv"))
```

Do the same with IUCN species maps.

```{r}
#iucn rasters
iucn_rasters <- list.files(file.path(dir_anx, "bio/spp_presence_rasters"), full.names = T)

#for IUCN, first remove all rasters we don't want. I know this is ugly!
iucn_list <- iucn_rasters[substr(iucn_rasters, 69, nchar(iucn_rasters)-4) %in% iucn_sids]
```

```{r, eval = F}
get_sp_rgn_iucn <- function(spp){
  
  # update the progress bar (tick()) and print progress (print())
  pb$tick()$print()
  #get SID (species ID)
  sid <- as.numeric(substr(spp, 69, nchar(spp)-4))
  
  #read in raster
  r <- raster(spp) %>%
    crop(nergns_rast)
  
  #use the raster::zonal() function to get a dataframe that lists each region where the species is found and a column with the species name
  z <- zonal(r, nergns_rast, na.rm=T, fun = "count") %>%
    as.data.frame() %>%
    rename(rgn_id = zone) %>%
    mutate(species = spp,
           iucn_sid = sid) %>%
    filter(count > 0) %>% 
    select(-count)
}

pb <- progress_estimated(length(iucn_list)) #progress bar
iucn_spp_rgns <- purrr::map_df(iucn_list, get_sp_rgn_iucn)

# add in rgn 12 (whole northeast)

rgn12_iucn <- iucn_spp_rgns %>%
  select(-rgn_id) %>%
  distinct() %>%
  mutate(rgn_id = 12)

iucn_spp_rgns_ne <- bind_rows(iucn_spp_rgns, rgn12_iucn)

write.csv(iucn_spp_rgns_ne, file = "data/6_ne_iucn_spp_by_rgn.csv")
```

This function maps each IUCN species presence/absence raster and overlays the cellIDs to return a dataframe of cellIDs where the species is found, and a column with SID (species ID)

```{r, eval = F}
sid_extract <- function(spp){
  
  # update the progress bar (tick()) and print progress (print())
  pb$tick()$print()
  #get SID (species ID)
  sid <- as.numeric(substr(spp, 69, nchar(spp)-4))
  
  #read in raster
  r <- raster(spp) %>%
    crop(nergns_rastke)
  
  #swap out 1 for SID simply by multiplying since we have 1 identifying presence (could also use subs but I think this is faster)
  s <- r*sid
  
  #use the raster::zonal() function to get a dataframe that contains cellids and SID as two columns
  z <- zonal(s, cellid_rast, na.rm=T) %>%
    as.data.frame() %>%
    filter(!is.na(mean)) %>%
    rename(cellID = zone,
           SID = mean) %>%
    inner_join(rgn_cells)
}

pb <- progress_estimated(length(iucn_list)) #progress bar
sid_cells <- purrr::map_df(iucn_list, sid_extract)

#save to mazu because of file size (2.3 GB!)
write.csv(sid_cells, file = file.path(dir_anx, "bio/iucn_sid_cells.csv"))
```

Combine dataportal and IUCN rgn area dataframes for a single one to save in our layers folder for `ne-scores`

```{r}
#get northeast species cleaned up
ne_scinames <- read_csv("data/4_ne_dataportal_spp_common_sci.csv") %>%
  select(-X1) %>%
  mutate(sciname = tolower(sciname))

dp_spp_rgns <- read_csv("data/6_ne_dataportal_spp_by_rgn.csv")

nedp <- dp_spp_rgns %>%
  mutate(sciname = str_replace_all(species, "_", " "),
         species = str_replace_all(sciname, "normalized AEA", ""),
         common  = trimws(tolower(species), "both")) %>%
  select(-sciname, -species) %>%
  left_join(ne_scinames)

#iucn
iucn_ne_spp_info <- read_csv("data/1_iucn_spp_in_ne.csv")

iucn_spp_rgns <- read_csv("data/6_ne_iucn_spp_by_rgn.csv")

iucn <- iucn_spp_rgns %>%
  left_join(iucn_ne_spp_info, by = "iucn_sid") %>%
  select(rgn_id, sciname, common = common_name) %>%
  mutate(sciname = tolower(sciname),
         common = tolower(common)) 

all_spp_rgn <- nedp %>%
  bind_rows(iucn) %>%
  mutate(year = 2017) %>% #we have to have at least one year for the OHI model calculation to work
  select(-X1)

write_csv(all_spp_rgn, "~/github/ne-scores/region/layers/spp_rgns.csv")
  
```
