==> prep/bio/spp/7_assign_spp_status_scores.Rmd <==
---
title: "Assign species status scores by region"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

# Summary

This script assigns each species & OHI Northeast region combo (e.g. Loggerhead sea turtle in Rhode Island) a score between 0 and 1 based on the conservation status.

# Setup

```{r setup, warning = F, message = F}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
source('~/github/ne-prep/src/R/common.R')
```

# Get data

Load the species info table that contains all species, the map source, the state/geo location in which it has an assessment and the assessment score (betwee 0 and 1).

```{r}
spp_status    <- read_csv("data/5_spp_status.csv") %>%
   select(-X1)
status_scores <- read_csv("data/natserv_status_scores.csv") %>% select(-X1)
```

This gives us a dataframe of all species and their specific status and scores. But what it doesn't tell us is where the species ranges are located. For example, we can see that Atlantic silverside has different statuses based on some state/geographic levels of assessment. But we don't know if Atlantic silverside is located in New Hampshire and Maine as well.

```{r}
filter(spp_status, common == "atlantic silverside")
```

We need to link species and regions to states in the `spp_status`.

It actually looks like all species have at least a USA designation. But there are some cases where the USA status is *NNR* (equivalent to NA) but IUCN has an actual status. For example:

```{r}
filter(spp_status, common == "atlantic bluefin tuna")
```



```{r}
expand_spp_status <- spp_status %>%
  left_join(rgn_data, by = c("state" = "state_abv")) %>%
  select(-area_km2, -trend, -state.y) %>%
  left_join(status_scores) %>%
  mutate(rgn_id = case_when(
    status_scale == "USA" ~ 12,  #assigning rgn_id numbers to USA and IUCN (12 and 13 respectively)
    status_scale == "IUCN" ~ 13,
    TRUE ~ as.numeric(rgn_id)
  )) %>%
  select(common, sciname, source, iucn_sid, status, status_scale, rgn_id, score) %>%
  group_by(common, sciname, source, iucn_sid) %>%
  complete(rgn_id = c(1:13)) %>% #had to expand through 13 so we don't lose the rows with IUCN and USA
  distinct() %>%
  ungroup() %>%
  left_join(rgn_data, by = "rgn_id") %>%
  select(-area_km2, -state, -state_abv)
```

Scenarios for gapfilling (if score == NA):
- take score from USA (rgn_id 12). If that is also NA
- take scores from IUCN (rgn_id 13)
- If USA and IUCN are NA but there are states with scores, take the average across those states.

Create species and geog status table. This tells us the IUCN level status for each species, and the US designated status for each species.
```{r}
iucn_usa_status_table <- expand_spp_status %>%
  select(common, sciname, status_scale, score) %>%
  filter(status_scale %in% c("IUCN", "USA")) %>%
  distinct() %>%
  group_by(common, sciname) %>% 
  mutate(count = n()) %>% #this should be 1 or 2. bigeye sand tiger has 3 entries... need to filter out
  ungroup() %>%
  filter(!(common == 'bigeye sand tiger' & status_scale == 'IUCN' & is.na(score) & count == 3),
         common != "bonaparte's gull") %>% #this is in twice. once as bonaparte's gull and also bonapartes gull. second is correct
  mutate(common = ifelse(
    common == "	g<fc>nther<92>s lanternfish", "gunther's lanternfish", common
  )) %>%
  tidyr::spread(status_scale, score)
```

Now we can assign scores to the species/regions that are NA in `expand_spp_status` table.

```{r}
full_sp_rgn_scores <- expand_spp_status %>%
  mutate(score1 = ifelse(is.na(score), iucn_usa_status_table$USA[match(sciname, iucn_usa_status_table$sciname)], score)) %>% #first if score is NA then find USA score in iucn_usa_status_table
  mutate(score2 = ifelse(is.na(score1), iucn_usa_status_table$IUCN[match(sciname, iucn_usa_status_table$sciname)], score1)) %>%
  select(-score, -score1) %>%
  rename(score = score2) %>%
  mutate(year = 2017) #have to have a year for OHI model to run... it will backfill other years
```

Let's look at the species that still have no score

```{r}
no_score <- full_sp_rgn_scores %>%
  filter(is.na(score))

#are there any species status assessments here? Or all data deficient or NA
unique(no_score$status)
```

All are NA. So we have every species scored that we can - at either the state, US or global (IUCN) level.

# Save list of all species and regions

```{r}
species_list <- full_sp_rgn_scores %>%
  mutate(`Common Name` = stringr::str_to_sentence(common),
         rgn_name = ifelse(rgn_id == 12, "Northeast", rgn_name)) %>% 
  select(`Common Name`, Scientific = sciname, Region = rgn_name) %>%
  distinct() %>%
  filter(!is.na(rgn_name))

write_csv(species_list, "data/species_list_by_rgn.csv")
```



# Save to toolbox

```{r}
write.csv(full_sp_rgn_scores, file = "~/github/ne-scores/region/layers/spp_status_scores.csv")
```
