==> prep/bio/spp/9_ecological_integrity_resilience_layer.Rmd <==
---
title: "Ecological integrity layer"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
source('~/github/ne-prep/src/R/common.R')
```

Read in the species status scores saved to the toolbox

```{r}
spp_scores <- read_csv("~/github/ne-scores/region/layers/spp_status_scores.csv") %>%
  mutate(sciname = tolower(sciname))
spp_rgns  <- read_csv("~/github/ne-scores/region/layers/spp_rgns.csv")
```

Calculate ecological integrity in the same way that we calculate status in the toolbox.

```{r}
#calculate status
spp_status <- spp_rgns %>%
    left_join(spp_scores, by = c("common", "sciname", "rgn_id", "year")) %>%
    select(rgn_id, common, sciname, score, year) %>%
    filter(!is.na(score)) %>%
    group_by(rgn_id, year) %>%
    mutate(mean_status = mean(score)) %>%
    distinct() %>%
    ungroup() %>%
    mutate(resilience_score = (0.75-mean_status)/0.75) %>%
    select(rgn_id, year, resilience_score)
```

# Save

```{r}
write_csv(spp_status, "~/github/ne-scores/region/layers/res_eco_integrity.csv")
```
